re-structure the modules
